e2e framework refactor
