+++ e2e-polybft/cardanofw/test-apex-system.go
@@ -1,307 +1,742 @@
 package cardanofw
 
 import (
+	"bytes"
 	"context"
+	"encoding/hex"
 	"errors"
 	"fmt"
+	"io"
+	"math/big"
+	"os"
+	"regexp"
+	"strings"
 	"sync"
 	"testing"
+	"time"
 
-	"github.com/0xPolygon/polygon-edge/consensus/polybft"
+	"github.com/0xPolygon/polygon-edge/consensus/polybft/wallet"
+	"github.com/0xPolygon/polygon-edge/contracts"
+	"github.com/0xPolygon/polygon-edge/crypto"
+	"github.com/0xPolygon/polygon-edge/e2e-polybft/framework"
+	"github.com/0xPolygon/polygon-edge/txrelayer"
 	"github.com/0xPolygon/polygon-edge/types"
 	cardanowallet "github.com/Ethernal-Tech/cardano-infrastructure/wallet"
 	"github.com/stretchr/testify/require"
+	"github.com/umbracle/ethgo"
 )
-
-type RunCardanoClusterConfig struct {
-	ID                 int
-	NodesCount         int
-	NetworkType        cardanowallet.CardanoNetworkType
-	InitialFundsKeys   []string
-	InitialFundsAmount uint64
-}
 
 type ApexSystem struct {
 	PrimeCluster  *TestCardanoCluster
 	VectorCluster *TestCardanoCluster
 	Nexus         *TestEVMBridge
-	Bridge        *TestCardanoBridge
+	BridgeCluster *framework.TestCluster
 
-	Config *ApexSystemConfig
+	validators  []*TestApexValidator
+	relayerNode *framework.Node
+
+	PrimeMultisigAddr     string
+	PrimeMultisigFeeAddr  string
+	VectorMultisigAddr    string
+	VectorMultisigFeeAddr string
+
+	relayerWallet *crypto.ECDSAKey
+	bladeAdmin    *crypto.ECDSAKey
+
+	config *ApexSystemConfig
 }
 
-type ApexSystemConfig struct {
-	// Apex
-	VectorEnabled bool
-
-	APIValidatorID int // -1 all validators
-	APIPortStart   int
-	APIKey         string
-
-	VectorTTLInc                uint64
-	VectorSlotRoundingThreshold uint64
-	PrimeTTLInc                 uint64
-	PrimeSlotRoundingThreshold  uint64
-
-	TelemetryConfig               string
-	TargetOneCardanoClusterServer bool
-
-	BladeValidatorCount int
-
-	PrimeClusterConfig  *RunCardanoClusterConfig
-	VectorClusterConfig *RunCardanoClusterConfig
-
-	// Nexus EVM
-	NexusEnabled bool
-
-	NexusValidatorCount   int
-	NexusStartingPort     int64
-	NexusBurnContractInfo *polybft.BurnContractInfo
-
-	CustomOracleHandler  func(mp map[string]interface{})
-	CustomRelayerHandler func(mp map[string]interface{})
-}
-
-type ApexSystemOptions func(*ApexSystemConfig)
-
-func WithAPIValidatorID(apiValidatorID int) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.APIValidatorID = apiValidatorID
-	}
-}
-
-func WithAPIPortStart(apiPortStart int) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.APIPortStart = apiPortStart
-	}
-}
-
-func WithAPIKey(apiKey string) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.APIKey = apiKey
-	}
-}
-
-func WithVectorTTL(threshold, ttlInc uint64) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.VectorSlotRoundingThreshold = threshold
-		h.VectorTTLInc = ttlInc
-	}
-}
-
-func WithPrimeTTL(threshold, ttlInc uint64) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.PrimeSlotRoundingThreshold = threshold
-		h.PrimeTTLInc = ttlInc
-	}
-}
-
-func WithTelemetryConfig(tc string) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.TelemetryConfig = tc // something like "0.0.0.0:5001,localhost:8126"
-	}
-}
-
-func WithTargetOneCardanoClusterServer(targetOneCardanoClusterServer bool) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.TargetOneCardanoClusterServer = targetOneCardanoClusterServer
-	}
-}
-
-func WithVectorEnabled(enabled bool) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.VectorEnabled = enabled
-	}
-}
-
-func WithNexusEnabled(enabled bool) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.NexusEnabled = enabled
-	}
-}
-
-func WithNexusStartintPort(port int64) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.NexusStartingPort = port
-	}
-}
-
-func WithNexusBurningContract(contractInfo *polybft.BurnContractInfo) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.NexusBurnContractInfo = contractInfo
-	}
-}
-
-func WithPrimeClusterConfig(config *RunCardanoClusterConfig) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.PrimeClusterConfig = config
-	}
-}
-
-func WithVectorClusterConfig(config *RunCardanoClusterConfig) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.VectorClusterConfig = config
-	}
-}
-
-func WithCustomConfigHandlers(callbackOracle, callbackRelayer func(mp map[string]interface{})) ApexSystemOptions {
-	return func(h *ApexSystemConfig) {
-		h.CustomOracleHandler = callbackOracle
-		h.CustomRelayerHandler = callbackRelayer
-	}
-}
-
-func (as *ApexSystemConfig) ServiceCount() int {
-	// Prime
-	count := 1
-
-	if as.VectorEnabled {
-		count++
-	}
-
-	if as.NexusEnabled {
-		count++
-	}
-
-	return count
-}
-
-func newApexSystemConfig(opts ...ApexSystemOptions) *ApexSystemConfig {
-	config := &ApexSystemConfig{
-		APIValidatorID: 1,
-		APIPortStart:   40000,
-		APIKey:         "test_api_key",
-
-		BladeValidatorCount: 4,
-
-		PrimeClusterConfig: &RunCardanoClusterConfig{
-			ID:          0,
-			NodesCount:  4,
-			NetworkType: cardanowallet.TestNetNetwork,
-		},
-		VectorClusterConfig: &RunCardanoClusterConfig{
-			ID:          1,
-			NodesCount:  4,
-			NetworkType: cardanowallet.VectorTestNetNetwork,
-		},
-
-		NexusValidatorCount: 4,
-		NexusStartingPort:   int64(30400),
-
-		VectorEnabled: true,
-		NexusEnabled:  false,
-		NexusBurnContractInfo: &polybft.BurnContractInfo{
-			BlockNumber: 0,
-			Address:     types.ZeroAddress,
-		},
-
-		CustomOracleHandler:  nil,
-		CustomRelayerHandler: nil,
-	}
-
+func NewApexSystem(
+	dataDirPath string, opts ...ApexSystemOptions,
+) *ApexSystem {
+	config := getDefaultApexSystemConfig()
 	for _, opt := range opts {
 		opt(config)
 	}
 
-	return config
+	validators := make([]*TestApexValidator, config.BladeValidatorCount)
+
+	for i := 0; i < config.BladeValidatorCount; i++ {
+		validators[i] = NewTestApexValidator(dataDirPath, i+1)
+	}
+
+	bridge := &ApexSystem{
+		validators: validators,
+		config:     config,
+	}
+
+	return bridge
 }
 
-func RunApexBridge(
-	t *testing.T, ctx context.Context,
-	apexOptions ...ApexSystemOptions,
-) *ApexSystem {
+func (a *ApexSystem) StopChains() {
+	wg := sync.WaitGroup{}
+	errorsContainer := [2]error{}
+
+	fmt.Println("Stopping chains...")
+
+	if a.PrimeCluster != nil {
+		wg.Add(1)
+
+		go func() {
+			defer wg.Done()
+
+			errorsContainer[0] = a.PrimeCluster.Stop()
+		}()
+	}
+
+	if a.VectorCluster != nil {
+		wg.Add(1)
+
+		go func() {
+			defer wg.Done()
+
+			errorsContainer[1] = a.VectorCluster.Stop()
+		}()
+	}
+
+	if a.Nexus != nil {
+		wg.Add(1)
+
+		go func() {
+			defer wg.Done()
+
+			a.Nexus.Cluster.Stop()
+		}()
+	}
+
+	if a.BridgeCluster != nil {
+		wg.Add(1)
+
+		go func() {
+			defer wg.Done()
+
+			fmt.Printf("Cleaning up apex bridge\n")
+			a.BridgeCluster.Stop()
+			fmt.Printf("Done cleaning up apex bridge\n")
+		}()
+	}
+
+	wg.Wait()
+
+	fmt.Printf("Chains has been stopped...%v\n", errors.Join(errorsContainer[:]...))
+}
+
+func (a *ApexSystem) StartChains(t *testing.T) {
 	t.Helper()
 
-	apexConfig := newApexSystemConfig(apexOptions...)
-	apexSystem := &ApexSystem{Config: apexConfig}
 	wg := &sync.WaitGroup{}
-	serviceCount := apexConfig.ServiceCount()
 	errorsContainer := [3]error{}
 
-	wg.Add(serviceCount)
+	wg.Add(a.config.ServiceCount())
 
 	go func() {
 		defer wg.Done()
 
-		apexSystem.PrimeCluster, errorsContainer[0] = RunCardanoCluster(t, ctx, apexConfig.PrimeClusterConfig)
+		a.PrimeCluster, errorsContainer[0] = RunCardanoCluster(t, a.config.PrimeClusterConfig)
 	}()
 
-	if apexConfig.VectorEnabled {
+	if a.config.VectorEnabled {
 		go func() {
 			defer wg.Done()
 
-			apexSystem.VectorCluster, errorsContainer[1] = RunCardanoCluster(t, ctx, apexConfig.VectorClusterConfig)
+			a.VectorCluster, errorsContainer[1] = RunCardanoCluster(t, a.config.VectorClusterConfig)
 		}()
 	}
 
-	if apexConfig.NexusEnabled {
+	if a.config.NexusEnabled {
 		go func() {
 			defer wg.Done()
 
-			apexSystem.Nexus, errorsContainer[2] = RunEVMChain(t, apexConfig)
+			a.Nexus, errorsContainer[2] = RunEVMChain(t, a.config)
 		}()
 	}
-
-	t.Cleanup(func() {
-		fmt.Println("Stopping chains...")
-
-		if apexSystem.PrimeCluster != nil {
-			wg.Add(1)
-
-			go func() {
-				defer wg.Done()
-
-				errorsContainer[0] = apexSystem.PrimeCluster.Stop()
-			}()
-		}
-
-		if apexSystem.VectorCluster != nil {
-			wg.Add(1)
-
-			go func() {
-				defer wg.Done()
-
-				errorsContainer[1] = apexSystem.VectorCluster.Stop()
-			}()
-		}
-
-		if apexSystem.Nexus != nil {
-			wg.Add(1)
-
-			go func() {
-				defer wg.Done()
-
-				apexSystem.Nexus.Cluster.Stop()
-			}()
-		}
-
-		if apexSystem.Bridge != nil {
-			wg.Add(1)
-
-			go func() {
-				defer wg.Done()
-
-				fmt.Printf("Cleaning up apex bridge\n")
-				apexSystem.Bridge.StopValidators()
-				fmt.Printf("Done cleaning up apex bridge\n")
-			}()
-		}
-
-		wg.Wait()
-
-		fmt.Printf("Chains has been stopped...%v\n", errors.Join(errorsContainer[:]...))
-	})
-
-	fmt.Println("Starting chains...")
 
 	wg.Wait()
 
+	t.Cleanup(a.StopChains) // cleanup everything
+
 	require.NoError(t, errors.Join(errorsContainer[:]...))
+}
 
-	fmt.Println("Chains have been started...")
+func (a *ApexSystem) StartBridgeChain(t *testing.T) {
+	t.Helper()
 
-	bridgeDataDir := "../../e2e-bridge-data-tmp-" + t.Name() // path.Join(path.Dir(primeCluster.Config.TmpDir), "bridge")
-	apexSystem.Bridge = SetupAndRunApexBridge(t, ctx, bridgeDataDir, apexSystem)
+	bladeAdmin, err := crypto.GenerateECDSAKey()
+	require.NoError(t, err)
 
-	return apexSystem
+	a.bladeAdmin = bladeAdmin
+	a.BridgeCluster = framework.NewTestCluster(t, a.config.BladeValidatorCount,
+		framework.WithBladeAdmin(bladeAdmin.Address().String()),
+	)
+
+	for idx, validator := range a.validators {
+		require.NoError(t, validator.SetClusterAndServer(a.BridgeCluster, a.BridgeCluster.Servers[idx]))
+	}
+
+	a.BridgeCluster.WaitForReady(t)
+}
+
+func (a *ApexSystem) GetPrimeGenesisWallet(t *testing.T) cardanowallet.IWallet {
+	t.Helper()
+
+	result, err := GetGenesisWalletFromCluster(a.PrimeCluster.Config.TmpDir, 2)
+	require.NoError(t, err)
+
+	return result
+}
+
+func (a *ApexSystem) GetVectorGenesisWallet(t *testing.T) cardanowallet.IWallet {
+	t.Helper()
+
+	result, err := GetGenesisWalletFromCluster(a.VectorCluster.Config.TmpDir, 2)
+	require.NoError(t, err)
+
+	return result
+}
+
+func (a *ApexSystem) GetPrimeTxProvider() cardanowallet.ITxProvider {
+	return cardanowallet.NewTxProviderOgmios(a.PrimeCluster.OgmiosURL())
+}
+
+func (a *ApexSystem) GetVectorTxProvider() cardanowallet.ITxProvider {
+	return cardanowallet.NewTxProviderOgmios(a.VectorCluster.OgmiosURL())
+}
+
+func (a *ApexSystem) CreateAndFundUser(t *testing.T, ctx context.Context, sendAmount uint64) *TestApexUser {
+	t.Helper()
+
+	user := NewTestApexUser(t, a.PrimeCluster.Config.NetworkType, a.GetVectorNetworkType())
+
+	txProviderPrime := a.GetPrimeTxProvider()
+	// Fund prime address
+	primeGenesisWallet := a.GetPrimeGenesisWallet(t)
+
+	user.SendToUser(
+		t, ctx, txProviderPrime, primeGenesisWallet, sendAmount, a.PrimeCluster.NetworkConfig())
+
+	fmt.Printf("Prime user address funded\n")
+
+	if a.config.VectorEnabled {
+		txProviderVector := a.GetVectorTxProvider()
+		// Fund vector address
+		vectorGenesisWallet := a.GetVectorGenesisWallet(t)
+
+		user.SendToUser(
+			t, ctx, txProviderVector, vectorGenesisWallet, sendAmount, a.VectorCluster.NetworkConfig())
+
+		fmt.Printf("Vector user address funded\n")
+	}
+
+	return user
+}
+
+func (a *ApexSystem) CreateAndFundExistingUser(
+	t *testing.T, ctx context.Context, primePrivateKey, vectorPrivateKey string, sendAmount uint64,
+	primeNetworkConfig TestCardanoNetworkConfig, vectorNetworkConfig TestCardanoNetworkConfig,
+) *TestApexUser {
+	t.Helper()
+
+	user := NewTestApexUserWithExistingWallets(t, primePrivateKey, vectorPrivateKey,
+		primeNetworkConfig.NetworkType, vectorNetworkConfig.NetworkType)
+
+	txProviderPrime := a.GetPrimeTxProvider()
+	txProviderVector := a.GetVectorTxProvider()
+
+	// Fund prime address
+	primeGenesisWallet := a.GetPrimeGenesisWallet(t)
+
+	user.SendToUser(t, ctx, txProviderPrime, primeGenesisWallet, sendAmount, primeNetworkConfig)
+
+	fmt.Printf("Prime user address funded\n")
+
+	// Fund vector address
+	vectorGenesisWallet := a.GetVectorGenesisWallet(t)
+
+	user.SendToUser(t, ctx, txProviderVector, vectorGenesisWallet, sendAmount, vectorNetworkConfig)
+
+	fmt.Printf("Vector user address funded\n")
+
+	return user
+}
+
+func (a *ApexSystem) CreateAndFundNexusUser(ctx context.Context, ethAmount uint64) (*wallet.Account, error) {
+	user, err := wallet.GenerateAccount()
+	if err != nil {
+		return nil, err
+	}
+
+	txRelayer, err := txrelayer.NewTxRelayer(txrelayer.WithClient(a.Nexus.Cluster.Servers[0].JSONRPC()))
+	if err != nil {
+		return nil, err
+	}
+
+	addr := user.Address()
+
+	receipt, err := txRelayer.SendTransaction(
+		types.NewTx(types.NewLegacyTx(
+			types.WithFrom(a.Nexus.Admin.Ecdsa.Address()),
+			types.WithTo(&addr),
+			types.WithValue(ethgo.Ether(ethAmount)),
+		)),
+		a.Nexus.Admin.Ecdsa)
+	if err != nil {
+		return nil, err
+	} else if receipt.Status != uint64(types.ReceiptSuccess) {
+		return nil, fmt.Errorf("fund user tx failed: %d", receipt.Status)
+	}
+
+	return user, nil
+}
+
+func (a *ApexSystem) GetVectorNetworkType() cardanowallet.CardanoNetworkType {
+	if a.config.VectorEnabled && a.VectorCluster != nil {
+		return a.VectorCluster.Config.NetworkType
+	}
+
+	return cardanowallet.TestNetNetwork // does not matter really
+}
+
+func (a *ApexSystem) GetBridgeAdmin() *crypto.ECDSAKey {
+	return a.bladeAdmin
+}
+
+func (a *ApexSystem) NexusCreateWalletsAndAddresses(createBLSKeys bool) (err error) {
+	if !a.config.NexusEnabled || len(a.validators) == 0 {
+		return nil
+	}
+
+	// relayer is on the first validator only
+	relayerValidator := a.validators[0]
+
+	if err = relayerValidator.createSpecificWallet("relayer-evm"); err != nil {
+		return err
+	}
+
+	a.relayerWallet, err = relayerValidator.getRelayerWallet()
+	if err != nil {
+		return err
+	}
+
+	for _, validator := range a.validators {
+		if createBLSKeys {
+			if err = validator.createSpecificWallet("batcher-evm"); err != nil {
+				return err
+			}
+		}
+
+		validator.BatcherBN256PrivateKey, err = validator.getBatcherWallet(!createBLSKeys)
+		if err != nil {
+			return err
+		}
+	}
+
+	return err
+}
+
+func (a *ApexSystem) GetRelayerWalletAddr() types.Address {
+	return a.relayerWallet.Address()
+}
+
+func (a *ApexSystem) GetValidatorsCount() int {
+	return len(a.validators)
+}
+
+func (a *ApexSystem) GetValidator(t *testing.T, idx int) *TestApexValidator {
+	t.Helper()
+
+	require.True(t, idx >= 0 && idx < len(a.validators))
+
+	return a.validators[idx]
+}
+
+func (a *ApexSystem) RegisterChains(fundTokenAmount uint64) error {
+	primeTokenSupply := new(big.Int).SetUint64(fundTokenAmount)
+	vectorTokenSupply := new(big.Int).SetUint64(fundTokenAmount)
+	nexusTokenSupplyDfm := new(big.Int).SetUint64(fundTokenAmount)
+	nexusTokenSupplyDfm.Mul(nexusTokenSupplyDfm, new(big.Int).Exp(big.NewInt(10), big.NewInt(6), nil))
+
+	errs := make([]error, len(a.validators))
+	wg := sync.WaitGroup{}
+
+	wg.Add(len(a.validators))
+
+	for i, validator := range a.validators {
+		go func(validator *TestApexValidator, indx int) {
+			defer wg.Done()
+
+			errs[indx] = validator.RegisterChain(ChainIDPrime, primeTokenSupply, ChainTypeCardano)
+			if errs[indx] != nil {
+				return
+			}
+
+			if a.config.VectorEnabled {
+				errs[indx] = validator.RegisterChain(ChainIDVector, vectorTokenSupply, ChainTypeCardano)
+				if errs[indx] != nil {
+					return
+				}
+			}
+
+			if a.config.NexusEnabled {
+				errs[indx] = validator.RegisterChain(ChainIDNexus, nexusTokenSupplyDfm, ChainTypeEVM)
+				if errs[indx] != nil {
+					return
+				}
+			}
+		}(validator, i)
+	}
+
+	wg.Wait()
+
+	return errors.Join(errs...)
+}
+
+func (a *ApexSystem) GenerateConfigs(
+	primeCluster *TestCardanoCluster,
+	vectorCluster *TestCardanoCluster,
+	nexus *TestEVMBridge,
+) error {
+	errs := make([]error, len(a.validators))
+	wg := sync.WaitGroup{}
+
+	wg.Add(len(a.validators))
+
+	for i, validator := range a.validators {
+		go func(validator *TestApexValidator, indx int) {
+			defer wg.Done()
+
+			telemetryConfig := ""
+			if indx == 0 {
+				telemetryConfig = a.config.TelemetryConfig
+			}
+
+			var (
+				primeNetworkAddr   string
+				primeNetworkMagic  uint
+				primeNetworkID     uint
+				vectorOgmiosURL    = "http://localhost:1000"
+				vectorNetworkAddr  = "localhost:5499"
+				vectorNetworkMagic = uint(0)
+				vectorNetworkID    = uint(0)
+				nexusNodeURL       = "http://localhost:5500"
+			)
+
+			if a.config.TargetOneCardanoClusterServer {
+				primeNetworkAddr = primeCluster.NetworkAddress()
+				primeNetworkMagic = primeCluster.Config.NetworkMagic
+				primeNetworkID = uint(primeCluster.Config.NetworkType)
+			} else {
+				primeServer := primeCluster.Servers[indx%len(primeCluster.Servers)]
+				primeNetworkAddr = primeServer.NetworkAddress()
+				primeNetworkMagic = primeServer.config.NetworkMagic
+				primeNetworkID = uint(primeServer.config.NetworkID)
+			}
+
+			if a.config.VectorEnabled {
+				vectorOgmiosURL = vectorCluster.OgmiosURL()
+
+				if a.config.TargetOneCardanoClusterServer {
+					vectorNetworkAddr = vectorCluster.NetworkAddress()
+					vectorNetworkMagic = vectorCluster.Config.NetworkMagic
+					vectorNetworkID = uint(vectorCluster.Config.NetworkType)
+				} else {
+					vectorServer := vectorCluster.Servers[indx%len(vectorCluster.Servers)]
+					vectorNetworkAddr = vectorServer.NetworkAddress()
+					vectorNetworkMagic = vectorServer.config.NetworkMagic
+					vectorNetworkID = uint(vectorServer.config.NetworkID)
+				}
+			}
+
+			if a.config.NexusEnabled {
+				nexusNodeURLIndx := 0
+				if a.config.TargetOneCardanoClusterServer {
+					nexusNodeURLIndx = indx % len(nexus.Cluster.Servers)
+				}
+
+				nexusNodeURL = nexus.Cluster.Servers[nexusNodeURLIndx].JSONRPCAddr()
+			}
+
+			errs[indx] = validator.GenerateConfigs(
+				primeNetworkAddr,
+				primeNetworkMagic,
+				primeNetworkID,
+				primeCluster.OgmiosURL(),
+				a.config.PrimeSlotRoundingThreshold,
+				a.config.PrimeTTLInc,
+				vectorNetworkAddr,
+				vectorNetworkMagic,
+				vectorNetworkID,
+				vectorOgmiosURL,
+				a.config.VectorSlotRoundingThreshold,
+				a.config.VectorTTLInc,
+				a.config.APIPortStart+indx,
+				a.config.APIKey,
+				telemetryConfig,
+				nexusNodeURL,
+			)
+		}(validator, i)
+	}
+
+	wg.Wait()
+
+	if err := errors.Join(errs...); err != nil {
+		return err
+	}
+
+	if handler := a.config.CustomOracleHandler; handler != nil {
+		for _, val := range a.validators {
+			err := UpdateJSONFile(
+				val.GetValidatorComponentsConfig(), val.GetValidatorComponentsConfig(), handler, false)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	if handler := a.config.CustomRelayerHandler; handler != nil {
+		for _, val := range a.validators {
+			if RunRelayerOnValidatorID != val.ID {
+				continue
+			}
+
+			if err := UpdateJSONFile(val.GetRelayerConfig(), val.GetRelayerConfig(), handler, false); err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+func (a *ApexSystem) StartValidatorComponents(ctx context.Context) (err error) {
+	for _, validator := range a.validators {
+		hasAPI := a.config.APIValidatorID == -1 || validator.ID == a.config.APIValidatorID
+
+		if err = validator.Start(ctx, hasAPI); err != nil {
+			return err
+		}
+	}
+
+	return err
+}
+
+func (a *ApexSystem) StartRelayer(ctx context.Context) (err error) {
+	for _, validator := range a.validators {
+		if RunRelayerOnValidatorID != validator.ID {
+			continue
+		}
+
+		a.relayerNode, err = framework.NewNodeWithContext(ctx, ResolveApexBridgeBinary(), []string{
+			"run-relayer",
+			"--config", validator.GetRelayerConfig(),
+		}, os.Stdout)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (a ApexSystem) StopRelayer() error {
+	if a.relayerNode == nil {
+		return errors.New("relayer not started")
+	}
+
+	return a.relayerNode.Stop()
+}
+
+func (a *ApexSystem) GetBridgingAPI() (string, error) {
+	apis, err := a.GetBridgingAPIs()
+	if err != nil {
+		return "", err
+	}
+
+	return apis[0], nil
+}
+
+func (a *ApexSystem) GetBridgingAPIs() (res []string, err error) {
+	for _, validator := range a.validators {
+		hasAPI := a.config.APIValidatorID == -1 || validator.ID == a.config.APIValidatorID
+
+		if hasAPI {
+			if validator.APIPort == 0 {
+				return nil, fmt.Errorf("api port not defined")
+			}
+
+			res = append(res, fmt.Sprintf("http://localhost:%d", validator.APIPort))
+		}
+	}
+
+	if len(res) == 0 {
+		return nil, fmt.Errorf("not running API")
+	}
+
+	return res, nil
+}
+
+func (a *ApexSystem) ApexBridgeProcessesRunning() bool {
+	if a.relayerNode == nil || a.relayerNode.ExitResult() != nil {
+		return false
+	}
+
+	for _, validator := range a.validators {
+		if validator.node == nil || validator.node.ExitResult() != nil {
+			return false
+		}
+	}
+
+	return true
+}
+
+func (a *ApexSystem) CreateCardanoWallets() (err error) {
+	for _, validator := range a.validators {
+		if err = validator.CardanoWalletCreate(ChainIDPrime); err != nil {
+			return err
+		}
+
+		if a.config.VectorEnabled {
+			if err = validator.CardanoWalletCreate(ChainIDVector); err != nil {
+				return err
+			}
+		}
+	}
+
+	return err
+}
+
+func (a *ApexSystem) FundCardanoMultisigAddresses(
+	ctx context.Context, fundTokenAmount uint64,
+) error {
+	const (
+		numOfRetries = 90
+		waitTime     = time.Second * 2
+	)
+
+	fund := func(cluster *TestCardanoCluster, fundTokenAmount uint64, addr string) (string, error) {
+		txProvider := cardanowallet.NewTxProviderOgmios(cluster.OgmiosURL())
+
+		defer txProvider.Dispose()
+
+		genesisWallet, err := GetGenesisWalletFromCluster(cluster.Config.TmpDir, 1)
+		if err != nil {
+			return "", err
+		}
+
+		txHash, err := SendTx(ctx, txProvider, genesisWallet, fundTokenAmount,
+			addr, cluster.NetworkConfig(), []byte{})
+		if err != nil {
+			return "", err
+		}
+
+		err = cardanowallet.WaitForTxHashInUtxos(
+			ctx, txProvider, addr, txHash, numOfRetries, waitTime, IsRecoverableError)
+		if err != nil {
+			return "", err
+		}
+
+		return txHash, nil
+	}
+
+	txHash, err := fund(a.PrimeCluster, fundTokenAmount, a.PrimeMultisigAddr)
+	if err != nil {
+		return err
+	}
+
+	fmt.Printf("Prime multisig addr funded: %s\n", txHash)
+
+	txHash, err = fund(a.PrimeCluster, fundTokenAmount, a.PrimeMultisigFeeAddr)
+	if err != nil {
+		return err
+	}
+
+	fmt.Printf("Prime fee addr funded: %s\n", txHash)
+
+	if a.config.VectorEnabled {
+		txHash, err := fund(a.VectorCluster, fundTokenAmount, a.VectorMultisigAddr)
+		if err != nil {
+			return err
+		}
+
+		fmt.Printf("Vector multisig addr funded: %s\n", txHash)
+
+		txHash, err = fund(a.VectorCluster, fundTokenAmount, a.VectorMultisigFeeAddr)
+		if err != nil {
+			return err
+		}
+
+		fmt.Printf("Vector fee addr funded: %s\n", txHash)
+	}
+
+	return nil
+}
+
+func (a *ApexSystem) CreateCardanoMultisigAddresse() (err error) {
+	a.PrimeMultisigAddr, a.PrimeMultisigFeeAddr, err = a.cardanoCreateAddress(a.PrimeCluster.Config.NetworkType, nil)
+	if err != nil {
+		return fmt.Errorf("failed to create addresses for prime: %w", err)
+	}
+
+	if a.config.VectorEnabled {
+		a.VectorMultisigAddr, a.VectorMultisigFeeAddr, err = a.cardanoCreateAddress(a.VectorCluster.Config.NetworkType, nil)
+		if err != nil {
+			return fmt.Errorf("failed to create addresses for vector: %w", err)
+		}
+	}
+
+	return nil
+}
+
+func (a *ApexSystem) cardanoCreateAddress(
+	network cardanowallet.CardanoNetworkType, keys []string,
+) (string, string, error) {
+	bridgeAdminPk, err := a.bladeAdmin.MarshallPrivateKey()
+	if err != nil {
+		return "", "", err
+	}
+
+	bothAddresses := false
+
+	args := []string{
+		"create-address",
+		"--network-id", fmt.Sprint(network),
+	}
+
+	if len(keys) == 0 {
+		args = append(args,
+			"--bridge-url", a.BridgeCluster.Servers[0].JSONRPCAddr(),
+			"--bridge-addr", contracts.Bridge.String(),
+			"--chain", GetNetworkName(network),
+			"--bridge-key", hex.EncodeToString(bridgeAdminPk))
+
+		bothAddresses = true
+	}
+
+	for _, key := range keys {
+		args = append(args, "--key", key)
+	}
+
+	var outb bytes.Buffer
+
+	err = RunCommand(ResolveApexBridgeBinary(), args, io.MultiWriter(os.Stdout, &outb))
+	if err != nil {
+		return "", "", err
+	}
+
+	if !bothAddresses {
+		return strings.TrimSpace(strings.ReplaceAll(outb.String(), "Address = ", "")), "", nil
+	}
+
+	multisig, fee, output := "", "", outb.String()
+	reGateway := regexp.MustCompile(`Multisig Address\s*=\s*([^\s]+)`)
+	reNativeTokenWallet := regexp.MustCompile(`Fee Payer Address\s*=\s*([^\s]+)`)
+
+	if match := reGateway.FindStringSubmatch(output); len(match) > 0 {
+		multisig = match[1]
+	}
+
+	if match := reNativeTokenWallet.FindStringSubmatch(output); len(match) > 0 {
+		fee = match[1]
+	}
+
+	return multisig, fee, nil
 }
