+++ e2e-polybft/cardanofw/test-apex-validator.go
@@ -36,7 +36,7 @@ type CardanoWallet struct {
 	MultisigFee *cardanoWallet.Wallet `json:"fee"`
 }
 
-type TestCardanoValidator struct {
+type TestApexValidator struct {
 	ID                     int
 	APIPort                int
 	dataDirPath            string
@@ -46,17 +46,17 @@ type TestCardanoValidator struct {
 	BatcherBN256PrivateKey *bn256.PrivateKey
 }
 
-func NewTestCardanoValidator(
+func NewTestApexValidator(
 	dataDirPath string,
 	id int,
-) *TestCardanoValidator {
-	return &TestCardanoValidator{
+) *TestApexValidator {
+	return &TestApexValidator{
 		dataDirPath: filepath.Join(dataDirPath, fmt.Sprintf("validator_%d", id)),
 		ID:          id,
 	}
 }
 
-func (cv *TestCardanoValidator) SetClusterAndServer(
+func (cv *TestApexValidator) SetClusterAndServer(
 	cluster *framework.TestCluster, server *framework.TestServer,
 ) error {
 	cv.cluster = cluster
@@ -86,23 +86,23 @@ func (cv *TestCardanoValidator) SetClusterAndServer(
 	return nil
 }
 
-func (cv *TestCardanoValidator) GetBridgingConfigsDir() string {
+func (cv *TestApexValidator) GetBridgingConfigsDir() string {
 	return filepath.Join(cv.dataDirPath, BridgingConfigsDir)
 }
 
-func (cv *TestCardanoValidator) GetValidatorComponentsConfig() string {
+func (cv *TestApexValidator) GetValidatorComponentsConfig() string {
 	return filepath.Join(cv.GetBridgingConfigsDir(), ValidatorComponentsConfigFileName)
 }
 
-func (cv *TestCardanoValidator) GetRelayerConfig() string {
+func (cv *TestApexValidator) GetRelayerConfig() string {
 	return filepath.Join(cv.GetBridgingConfigsDir(), RelayerConfigFileName)
 }
 
-func (cv *TestCardanoValidator) GetNexusTestDir() string {
+func (cv *TestApexValidator) GetNexusTestDir() string {
 	return filepath.Join(cv.dataDirPath, NexusDir)
 }
 
-func (cv *TestCardanoValidator) CardanoWalletCreate(chainID string) error {
+func (cv *TestApexValidator) CardanoWalletCreate(chainID string) error {
 	return RunCommand(ResolveApexBridgeBinary(), []string{
 		"wallet-create",
 		"--chain", chainID,
@@ -110,7 +110,7 @@ func (cv *TestCardanoValidator) CardanoWalletCreate(chainID string) error {
 	}, os.Stdout)
 }
 
-func (cv *TestCardanoValidator) GetCardanoWallet(chainID string) (*CardanoWallet, error) {
+func (cv *TestApexValidator) GetCardanoWallet(chainID string) (*CardanoWallet, error) {
 	secretsMngr, err := cv.getSecretsManager(cv.dataDirPath)
 	if err != nil {
 		return nil, fmt.Errorf("failed to load wallet: %w", err)
@@ -132,7 +132,7 @@ func (cv *TestCardanoValidator) GetCardanoWallet(chainID string) (*CardanoWallet
 	return cardanoWallet, nil
 }
 
-func (cv *TestCardanoValidator) RegisterChain(
+func (cv *TestApexValidator) RegisterChain(
 	chainID string,
 	tokenSupply *big.Int,
 	chainType uint8,
@@ -148,7 +148,7 @@ func (cv *TestCardanoValidator) RegisterChain(
 	}, os.Stdout)
 }
 
-func (cv *TestCardanoValidator) GenerateConfigs(
+func (cv *TestApexValidator) GenerateConfigs(
 	primeNetworkAddress string,
 	primeNetworkMagic uint,
 	primeNetworkID uint,
@@ -227,7 +227,7 @@ func (cv *TestCardanoValidator) GenerateConfigs(
 	return common.CreateDirSafe(dbsPath, 0770)
 }
 
-func (cv *TestCardanoValidator) Start(ctx context.Context, runAPI bool) (err error) {
+func (cv *TestApexValidator) Start(ctx context.Context, runAPI bool) (err error) {
 	args := []string{
 		"run-validator-components",
 		"--config", cv.GetValidatorComponentsConfig(),
@@ -242,7 +242,7 @@ func (cv *TestCardanoValidator) Start(ctx context.Context, runAPI bool) (err err
 	return err
 }
 
-func (cv *TestCardanoValidator) Stop() error {
+func (cv *TestApexValidator) Stop() error {
 	if cv.node == nil {
 		return errors.New("validator not started")
 	}
@@ -250,7 +250,7 @@ func (cv *TestCardanoValidator) Stop() error {
 	return cv.node.Stop()
 }
 
-func (cv *TestCardanoValidator) createSpecificWallet(walletType string) error {
+func (cv *TestApexValidator) createSpecificWallet(walletType string) error {
 	return RunCommand(ResolveApexBridgeBinary(), []string{
 		"wallet-create",
 		"--chain", ChainIDNexus,
@@ -259,7 +259,7 @@ func (cv *TestCardanoValidator) createSpecificWallet(walletType string) error {
 	}, os.Stdout)
 }
 
-func (cv *TestCardanoValidator) getBatcherWallet(loadFromBlade bool) (*bn256.PrivateKey, error) {
+func (cv *TestApexValidator) getBatcherWallet(loadFromBlade bool) (*bn256.PrivateKey, error) {
 	secretsMngr, err := cv.getSecretsManager(cv.server.DataDir())
 	if err != nil {
 		return nil, fmt.Errorf("failed to load wallet: %w", err)
@@ -286,7 +286,7 @@ func (cv *TestCardanoValidator) getBatcherWallet(loadFromBlade bool) (*bn256.Pri
 	return bn256, nil
 }
 
-func (cv *TestCardanoValidator) getRelayerWallet() (*crypto.ECDSAKey, error) {
+func (cv *TestApexValidator) getRelayerWallet() (*crypto.ECDSAKey, error) {
 	secretsMngr, err := cv.getSecretsManager(cv.server.DataDir())
 	if err != nil {
 		return nil, fmt.Errorf("failed to load wallet: %w", err)
@@ -312,7 +312,7 @@ func (cv *TestCardanoValidator) getRelayerWallet() (*crypto.ECDSAKey, error) {
 	return pk, nil
 }
 
-func (cv *TestCardanoValidator) getSecretsManager(path string) (secretsCardano.SecretsManager, error) {
+func (cv *TestApexValidator) getSecretsManager(path string) (secretsCardano.SecretsManager, error) {
 	return secretsHelper.CreateSecretsManager(&secretsCardano.SecretsManagerConfig{
 		Path: path,
 		Type: secretsCardano.Local,
